backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
requests
beautifulsoup4
python-docx
pypdf
python-multipart